workers/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
nats-py==2.4.0
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# ML/AI dependencies
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytesseract==0.3.10
pillow==10.1.0

# Graph processing
networkx==3.2.1
neo4j==5.15.0

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0
